engine: extract esm/layers/grid_layer.js
