+++ engine/esm/index.js
@@ -225,6 +225,8 @@ export { Layer, DomainValue, AltUnits, FadeType } from "./layers/layer.js";
 // To keep API compatibility, we can't fix this typo.
 export { GreatCirlceRouteLayer } from "./layers/great_circle_route_layer.js";
 
+export { GridLayer } from "./layers/grid_layer.js";
+
 export {
     PointType,
     FolderGroup,
@@ -312,7 +314,6 @@ export {
     WWTControl,
     WWTControlBuilder,
     WWTElementEvent,
-    GridLayer,
     ImageSetLayer,
     Object3dLayer,
     Object3dLayerUI,
